fix: post merge conflicts resolution
