+++ Fiero/Views/Login/Recover Password/InputConfirmationCodeView.swift
@@ -1,5 +1,5 @@
 //
-//  InputConfirmationCode.swift
+//  InputConfirmationCodeView.swift
 //  Fiero
 //
 //  Created by João Gabriel Biazus de Quevedo on 07/10/22.
@@ -11,22 +11,21 @@ struct InputConfirmationCodeView: View {
     @Environment(\.presentationMode) var presentationMode
     
     @EnvironmentObject var userViewModel: UserViewModel
-    
+
     @State private var confirmationCode: String = ""
+    @State private var password: String = ""
+    @State private var confirmPassword: String = ""
+    
+    @State private var isPresentingErrorAlert: Bool = false
     @State var isPresentingNewPassword: Bool = false
     
-    
-    var textFont: Font {
-        return Tokens.FontStyle.callout.font()
-    }
-    var color: Color {
-        return Tokens.Colors.Neutral.High.pure.value
-    }
-    
+    @Binding var email: String
+
     var body: some View {
         ZStack {
             Tokens.Colors.Background.dark.value.ignoresSafeArea()
-            ScrollView{
+            
+            ScrollView {
                 VStack (spacing: Tokens.Spacing.xxxs.value){
                     Spacer()
                     Text("verificationcationCodeTitleLabel")
@@ -54,7 +53,7 @@ struct InputConfirmationCodeView: View {
                         isSecure: false,
                         isLowCase: true ,
                         isWrong: .constant(false),
-                        text: self.$confirmationCode)
+                        text: self.$password)
                     
                     CustomTextFieldView(
                         type: .none,
@@ -64,32 +63,94 @@ struct InputConfirmationCodeView: View {
                         isSecure: false,
                         isLowCase: true ,
                         isWrong: .constant(false),
-                        text: self.$confirmationCode)
+                        text: self.$confirmPassword)
                     
-                    ButtonComponent(
-                        style: .primary(isEnabled: true),
-                        text: "changePasswordButtonLabel",
-                        action: {
-                            //TODO: - verify confirmation code before go to next step
-                            self.isPresentingNewPassword.toggle()
-                        })
-                    ButtonComponent(
-                        style: .black(isEnabled: true),
-                        text: "resendEmailVerification",
-                        action: {
-                            //TODO: - verify confirmation code before go to next step
-                            
-                        })
-                    
+                    if userViewModel.keyboardShown {
+                        ButtonComponent(
+                            style: .primary(isEnabled: true),
+                            text: "changePasswordButtonLabel",
+                            action: {
+                                validateInput(password: self.password,
+                                                      confirmPassword: self.confirmPassword,
+                                                      validationCode: self.confirmationCode)
+                            })
+                        .padding(.bottom, Tokens.Spacing.sm.value)
+                    } else {
+                        ButtonComponent(
+                            style: .primary(isEnabled: true),
+                            text: "changePasswordButtonLabel",
+                            action: {
+                                validateInput(password: self.password,
+                                                    confirmPassword: self.confirmPassword,
+                                                    validationCode: self.confirmationCode)
+                            })
+                    }
                 }.padding(.horizontal,Tokens.Spacing.defaultMargin.value)
             }
         }
+        .alert(isPresented: self.$isPresentingErrorAlert, content: {
+            switch self.userViewModel.recoveryAccountSecondStepErrorCases {
+                case .none:
+                    return Alert(title: Text(RecoveryAccountSecondStepErrorCases.none.title),
+                                 message: Text(RecoveryAccountSecondStepErrorCases.none.message),
+                                 dismissButton: .cancel(Text("OK"), action: {
+                        self.isPresentingErrorAlert = false
+                        RootViewController.popToRootViewController()
+                    }))
+                case .emptyFields:
+                    return Alert(title: Text(RecoveryAccountSecondStepErrorCases.emptyFields.title),
+                                 message: Text(RecoveryAccountSecondStepErrorCases.emptyFields.message),
+                                 dismissButton: .cancel(Text("OK"), action: {
+                        self.isPresentingErrorAlert = false
+                    }))
+                case .wrongCode:
+                    return Alert(title: Text(RecoveryAccountSecondStepErrorCases.wrongCode.title),
+                                        message: Text(RecoveryAccountSecondStepErrorCases.wrongCode.message),
+                                        primaryButton: .default(Text(RecoveryAccountSecondStepErrorCases.wrongCode.primaryButton), action: {
+                                self.userViewModel.sendVerificationCode(for: self.email)
+                                self.isPresentingErrorAlert = false
+                           }),
+                                        secondaryButton: .destructive(Text(RecoveryAccountSecondStepErrorCases.wrongCode.secondaryButton), action: {
+                               self.isPresentingErrorAlert = false
+                                RootViewController.popToRootViewController()
+                           }))
+                case .unmatchedPasswords:
+                    return Alert(title: Text(RecoveryAccountSecondStepErrorCases.unmatchedPasswords.title),
+                                 message: Text(RecoveryAccountSecondStepErrorCases.unmatchedPasswords.message),
+                                 dismissButton: .cancel(Text(RecoveryAccountSecondStepErrorCases.unmatchedPasswords.primaryButton), action: {
+                        self.isPresentingErrorAlert = false
+                    }))
+                case .internalServerError:
+                    return Alert(title: Text(RecoveryAccountSecondStepErrorCases.internalServerError.title),
+                                 message: Text(RecoveryAccountSecondStepErrorCases.internalServerError.message),
+                                 dismissButton: .cancel(Text("OK"), action: {
+                        self.isPresentingErrorAlert = false
+                    }))
+            }
+        })
         .navigationBarHidden(true)
+    }
+    
+    func validateInput(password: String, confirmPassword: String, validationCode: String) {
+        if self.confirmationCode.isEmpty || self.password.isEmpty || self.confirmPassword.isEmpty {
+            self.userViewModel.recoveryAccountSecondStepErrorCases = .emptyFields
+            self.isPresentingErrorAlert = true
+        }
+        else {
+            if self.password != self.confirmPassword {
+                self.userViewModel.recoveryAccountSecondStepErrorCases = .unmatchedPasswords
+                self.isPresentingErrorAlert = true
+            }
+            else {
+                self.userViewModel.resetAccountPassword(with: self.confirmPassword, using: self.confirmationCode)
+                self.isPresentingErrorAlert = true
+            }
+        }
     }
 }
 
 struct InputConfirmationCode_Previews: PreviewProvider {
     static var previews: some View {
-        InputConfirmationCodeView()
+        InputConfirmationCodeView(email: .constant("flemis"))
     }
 }
